+++ packages/shadergradient-v2/src/types.ts
@@ -28,6 +28,9 @@ export type MeshT = {
   range?: 'enabled' | 'disabled' | string
   rangeStart?: number
   rangeEnd?: number
+  // Loop controls
+  loop?: 'on' | 'off'
+  loopDuration?: number
   positionX?: number
   positionY?: number
   positionZ?: number
@@ -77,6 +80,7 @@ export type GradientT = MeshT & {
   hoverState?: string
 
   enableTransition?: boolean
+  enableCameraUpdate?: boolean
 
   urlString?: string