+++ packages/ui/src/components/Shared/ShaderGradientStateless/useControlValues.ts
@@ -17,6 +17,8 @@ export function useControlValues(
   const [range] = useQueryState('range')
   const [rangeStart] = useQueryState('rangeStart')
   const [rangeEnd] = useQueryState('rangeEnd')
+  const [loop] = useQueryState('loop')
+  const [loopDuration] = useQueryState('loopDuration')
   const [uTime] = useQueryState('uTime')
   const [uSpeed] = useQueryState('uSpeed')
   const [uStrength] = useQueryState('uStrength')
@@ -69,6 +71,8 @@ export function useControlValues(
     range,
     rangeStart,
     rangeEnd,
+    loop,
+    loopDuration,
     frameRate,
     destination,
     format,
@@ -105,7 +109,7 @@ export function useControlValues(
   }
 
   if (control === 'props')
-    return formatFramerProps(clean({ ...queryProps, ...props }) as any)
+    return clean({ ...queryProps, ...formatFramerProps(props) }) as any
   else if (control === 'query')
     return clean(
       urlString