+++ apps/example-nextjs-dev/CHANGELOG.md
@@ -1,5 +1,33 @@
 # example-nextjs-dev
 
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies
+  - @shadergradient/react@2.4.5
+
+## 0.1.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @shadergradient/react@2.4.4
+
+## 0.1.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @shadergradient/react@2.4.1
+
+## 0.1.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @shadergradient/react@2.4.0
+
 ## 0.1.48
 
 ### Patch Changes