+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,119 @@
 # ui
 
+## 0.1.39
+
+### Patch Changes
+
+- fix override
+
+## 0.1.38
+
+### Patch Changes
+
+- update HideBanner override
+
+## 0.1.37
+
+### Patch Changes
+
+- fix formatFramerProps
+
+## 0.1.36
+
+### Patch Changes
+
+- InitCustomize
+
+## 0.1.35
+
+### Patch Changes
+
+- withClickOutsideClosed
+
+## 0.1.34
+
+### Patch Changes
+
+- improve TextAnimation
+
+## 0.1.33
+
+### Patch Changes
+
+- restore
+
+## 0.1.30
+
+### Patch Changes
+
+- enableFadeOut
+
+## 0.1.29
+
+### Patch Changes
+
+- ScrollableTextBox property controls
+
+## 0.1.28
+
+### Patch Changes
+
+- update ScrollableTextBox anim
+
+## 0.1.27
+
+### Patch Changes
+
+- refactor TextAnimation component
+
+## 0.1.26
+
+### Patch Changes
+
+- replace framer's createStore with zustand
+
+## 0.1.25
+
+### Patch Changes
+
+- restore ssr fixes
+
+## 0.1.21
+
+### Patch Changes
+
+- user-select style
+
+## 0.1.20
+
+### Patch Changes
+
+- fix styles
+
+## 0.1.19
+
+### Patch Changes
+
+- update ui for figma-plugin
+
+## 0.1.18
+
+### Patch Changes
+
+- zustand useScrollableBoxStore
+
+## 0.1.17
+
+### Patch Changes
+
+- improve ScrollableTextBox
+
+## 0.1.16
+
+### Patch Changes
+
+- fpsCount override
+
 ## 0.1.15
 
 ### Patch Changes